+++ src/main/java/com/DBP/ticketing_backend/domain/auth/dto/response/LoginResponseDto.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LoginResponseDto {
+    private Long userId;
     private String email;
     private String name;
     private String role;
