+++ src/main/java/com/DBP/ticketing_backend/global/config/SecurityConfig.java
@@ -4,6 +4,7 @@ import com.DBP.ticketing_backend.domain.auth.exception.JwtAccessDeniedHandler;
 import com.DBP.ticketing_backend.domain.auth.exception.JwtAuthenticationEntryPoint;
 import com.DBP.ticketing_backend.domain.auth.service.JwtAuthenticationFilter;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.context.annotation.Bean;
@@ -12,6 +13,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
 
 @Configuration
 @RequiredArgsConstructor
@@ -23,7 +25,15 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         return http.csrf(csrf -> csrf.disable())
-                .cors(cors -> cors.disable())
+                .cors(cors -> cors.configurationSource(request -> {
+                    CorsConfiguration config = new CorsConfiguration();
+                    config.setAllowedOrigins(List.of("http://localhost:3000"));
+                    config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+                    config.setAllowedHeaders(List.of("*"));
+                    config.setExposedHeaders(List.of("Authorization"));
+                    config.setAllowCredentials(true);
+                    return config;
+                }))
                 .sessionManagement(
                         session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .formLogin(form -> form.disable())
@@ -51,7 +61,8 @@ public class SecurityConfig {
                                             // 카카오페이 콜백 URL
                                             "/api/payment/success",
                                             "/api/payment/cancel",
-                                            "/api/payment/fail")
+                                            "/api/payment/fail",
+                                            "/api/place/**")
                                     .permitAll();
 
                             // 로그아웃은 인증 필요
@@ -59,7 +70,6 @@ public class SecurityConfig {
 
                             // ADMIN만 접근 가능
                             auth.requestMatchers("/api/admin/**").hasRole("ADMIN");
-                            auth.requestMatchers("/api/place/**").hasRole("ADMIN");
 
                             // HOST만 접근 가능
                             auth.requestMatchers("/api/host/**").hasRole("HOST");
